🚧✨ use cookies to load item statuses and auth tokens
This has some pretty major issues right now. We make a lot of large network requests very quickly, especially on the story page. I'm not sure if it's possible to fix this without Hacker News allowing for fetching of only a specific item and not its children. Even then, this will be far from efficient 😕